update row counting/limiting to allow values exceeding 2^32 and adjust corresponding limits in cql SELECT requests
Currently, we cannot select more than 2^32 rows from a table, because when counting the number of rows
we are using 4-byte integers as variables holding the numbers and std::numeric_limits<uint32_t>::max()
as the maximum number of rows. This patch changes these types to 8-byte integers. In order to be able to
process queries originating from older nodes, the top 32 bits of new integers are optional and are
passed a the end of a serialized class - if they're not present we assume 0.

Fixes #5101.

+++ cql3/statements/select_statement.hh
@@ -148,11 +148,11 @@ public:
     bool has_group_by() const { return _group_by_cell_indices && !_group_by_cell_indices->empty(); }
 
 protected:
-    uint32_t do_get_limit(const query_options& options, ::shared_ptr<term> limit) const;
-    uint32_t get_limit(const query_options& options) const {
+    uint64_t do_get_limit(const query_options& options, ::shared_ptr<term> limit) const;
+    uint64_t get_limit(const query_options& options) const {
         return do_get_limit(options, _limit);
     }
-    uint32_t get_per_partition_limit(const query_options& options) const {
+    uint64_t get_per_partition_limit(const query_options& options) const {
         return do_get_limit(options, _per_partition_limit);
     }
     bool needs_post_query_ordering() const;
